Migrated tests to Mocha for backend
